+++ src/level/BspRayTracer.cpp
@@ -15,6 +15,9 @@
 
 #include "BspRayTracer.hpp"
 #include "entity/component/BrushModelComponent.hpp"
+#include "entity/Entity.hpp"
+#include "entity/system/BrushModelSystem.hpp"
+#include "engine/Engine.hpp"
 
 static fp calculateIntersectionT(fp startDist, fp endDist)
 {
@@ -143,18 +146,22 @@ bool BspRayTracer::visitNode(int nodeId, RayPoint& start, RayPoint& end)
     return exploreBothSidesOfNode(nodeId, start, end, intersectionT, startDist);
 }
 
-bool BspRayTracer::traceModel(BrushModelComponent& brushModel)
+bool BspRayTracer::traceModel(Entity* entityWithModel)
 {
-// FIXME: 2-20-2019
-#if false
-    currentModel = brushModel.model;
-    currentModelOwner = brushModel.owner;
-#endif
+    BrushModelComponent* brushModelComponent = entityWithModel->getComponent<BrushModelComponent>();
 
-    RayPoint start(ray.v[0] - brushModel.model->center, 0);
-    RayPoint end(ray.v[1] - brushModel.model->center, fp::fromInt(1));
+    if(brushModelComponent == nullptr)
+    {
+        x_system_error("Null brush model component\n");
+    }
 
-    bool hitSomething = !visitNode(getRootNode(*brushModel.model), start, end);
+    currentModel = brushModelComponent->model;
+    currentModelOwner = entityWithModel;
+
+    RayPoint start(ray.v[0] - currentModel->center, 0);
+    RayPoint end(ray.v[1] - currentModel->center, fp::fromInt(1));
+
+    bool hitSomething = !visitNode(getRootNode(*currentModel), start, end);
     
     return hitSomething;
 }
@@ -163,26 +170,21 @@ bool BspRayTracer::trace()
 {
     collision.location.t = maxValue<fp>();
 
-// FIXME: 2-20-2019
-#if false
+
     // FIXME: this could very easily be optimized
-    
-    auto brushModels = BrushModelComponent::getAll();
+
+    BrushModelSystem* brushModelSystem = Engine::getInstance()->brushModelSystem;
+
+    auto& brushModels = brushModelSystem->getAllEntities();
     
     bool hitSomething = false;
     
-    for(auto& brushModel : brushModels)
+    for(Entity* entity : brushModels)
     {
-        if(brushModel.model != nullptr)
-        {
-            hitSomething |= traceModel(brushModel);
-        }
+        hitSomething |= traceModel(entity);
     }
     
     return hitSomething;
-#endif
-
-    return false;
 }