+++ changelog.md
@@ -1,3 +1,8 @@
+### 3-21-2019
+* Genericized the various component systems into a single class
+* Fixed input handling by adding `InputSystem`
+* Fixed collision handling
+
 ### 3-20-2019
 #### Goals
 * Restore player movement